requirements.txt
python-dotenv==1.0.1
Flask==3.0.0
Werkzeug==3.0.1
